Refactor the deparsing of a CREATE EXTENSION to prevent NULL POINTER dereferences
During the static analysis project on the codebase this code has been flagged as having the potential for a null pointer dereference. Funily enough the author had already made a commment of it in the code this was not possible due to us setting the schema name before we pass in the statement. If we want to reuse this code in a later setting this comment might not always apply and we could actually run into null pointer dereference.

This patch changes a bit of the code arround to first of all make sure there is no NULL pointer dereference in this code anymore.
Secondly we allow for better deparsing by setting and adhering to the `if_not_exists` flag on the statement.
And finally add support for all syntax described in the documentation of postgres (FROM was missing).